GVT-3146 (part 2): Move publication comparison pair getter to publication service
